1/3 Saving SharedStringTable through streaming: Move the code for writing shared string table to a separate writer. Part of a move to 2 phase saving and DOM-less saving though streaming of elements.
